\name{build.coef.df}
\alias{build.coef.df}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Coefficient dataframe
}
\description{
Collects up coefficients for comparison across a bunch of models
}
\usage{
build.coef.df(fitset)
}

\arguments{
  \item{fitset}{
set of models
}
}

\value{
data.frame
}

\author{
ME
}
