\name{HelpfulFunctions-package}
\alias{HelpfulFunctions-package}
\alias{HelpfulFunctions}
\docType{package}
\title{
Helpful functions
}
\description{
A collection of functions I'd like to save
}
\details{
\tabular{ll}{
Package: \tab HelpfulFunctions\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-15\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
ME
}
