\name{last}
\alias{last}
\title{
  last element of a vector
}
\description{
Real simple, but just makes things a little prettier
}
\usage{
last(x, na.rm = T)
}
\arguments{
  \item{x}{
a vector
}
  \item{na.rm}{
if things end with a NA, should it be that or should it be the last real value?
}}

\value{
last element of x
}
\author{
Martha
}
