\name{lsos}
\alias{lsos}

\title{
Nicer list of objects in memory
}
\description{
http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session
}
\usage{
lsos(..., n = 10)
}

\arguments{
  \item{\dots}{
not used
}
  \item{n}{
number of objects to return
}
}

\value{
a list of loaded objects sorted by size
}
