\name{r2.corr.merPOP}
\alias{r2.corr.merPOP}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Coefficient of determiniation for (G)LMMs
}
\description{
Goodness of fit for (G)LMM based on population mean only - Marginal
}
\usage{
r2.corr.merPOP(m)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{m}{
 (G)LMM object
}
}
\details{
Correlation between responses and fitted values.  Marginal = based on population means, fixed effects only.
}
\value{
R2
}
\references{
glmm.wikidot.com/faq
}
